+++ app/console/components/header.tsx
@@ -25,8 +25,8 @@ const Header: React.FC = () => {
       setIsLoggedIn(JSON.parse(_tmp).val)
     }
   }, []);
-  return <div style={useMount.mount ? useMount.mountedStyle : useMount.unmountedStyle}>
-    <div className="sticky bg-background h-16 box shadow-lg flex items-center">
+  return <div className="sticky w-full h-16 top-0 left-0" style={useMount.mount ? useMount.mountedStyle : useMount.unmountedStyle}>
+    <div className="bg-background h-full box shadow-lg flex items-center">
       <>{useMount.ContentOrSkeleton}</>
     </div>
   </div>;