+++ CHANGELOG.md
@@ -17,13 +17,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Navigation: Side sheet Menu -> side bar [2e3daac9](https://github.com/arjunmnath/shopapp-website/commit/2e3daac9618f2bd7f820ef40c14240e31b8e9835) ✅
 - Header mount animation: Hardcode -> useMountAnimation hook [38b10944](https://github.com/arjunmnath/shopapp-website/commit/38b1094426b9a69e9041a940ac24cc419b675307) ✅
-- New Sale Page -> Completion 🔴
+- New Sale Page -> UI Completion ✅ [8c9f094](https://github.com/arjunmnath/shopapp-website/commit/8c9f094b36dec606a9995df67873a3936e3e49c9)
 
 ### fix
 
 - Header Skeleton misalignment [9fd46ee](https://github.com/arjunmnath/shopapp-website/commit/9fd46ee42d4439fb9a74d94f73dba8aeec8f3003) ✅
 - Bug#1 Phone Number type mismatch on client and server side 🟠
-
+- Issue#1 Header not showing Sticky Behavior 🔴
 ## [1.0.0] - 2023-08-15 (Unreleased)
 
 Here we would have the update steps for 1.2.4 for people to follow.